+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/ExperimentalBetaApi.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.core
 
 /**
  * The annotated method or class is in beta. It's public API may change or be removed in future releases.
