+++ ShopperInsights/src/main/java/com/braintreepayments/api/shopperinsights/EligiblePaymentsApiRequest.kt
@@ -1,6 +1,6 @@
 package com.braintreepayments.api.shopperinsights
 
-import com.braintreepayments.api.ExperimentalBetaApi
+import com.braintreepayments.api.core.ExperimentalBetaApi
 import org.json.JSONArray
 import org.json.JSONObject
